Create systems for cards

+++ Assets/_Project/Scripts/Enemy/EnemySpawnerDataSO.cs
@@ -13,7 +13,7 @@ public class EnemySpawnerDataSO : ScriptableObject
 public class Wave
 {
     public List<WaveData> waveDatas = new();
-    public int spawnRate;
+    public float spawnRate;
 }
 
 [Serializable]
